+++ packages/xrpl/src/models/transactions/delegateSet.ts
@@ -15,6 +15,11 @@ const NON_DELEGATABLE_TRANSACTIONS = new Set([
   'SignerListSet',
   'DelegateSet',
   'AccountDelete',
+  'Batch',
+  // Pseudo transactions below:
+  'EnableAmendment',
+  'SetFee',
+  'UNLModify',
 ])
 
 export interface Permission {