+++ packages/xrpl/src/sugar/autofill.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines -- lots of helper functions needed for autofill */
 import BigNumber from 'bignumber.js'
 import { xAddressToClassicAddress, isValidXAddress } from 'ripple-address-codec'
 
@@ -9,6 +10,7 @@ import {
   BaseTransaction,
   Payment,
   Transaction,
+  Batch,
 } from '../models'
 import { xrpToDrops } from '../utils'
 
@@ -224,6 +226,20 @@ function convertToClassicAddress<T extends BaseTransaction>(
   }
 }
 
+// Helper function to get the next valid sequence number for an account.
+async function getNextValidSequenceNumber(
+  client: Client,
+  account: string,
+): Promise<number> {
+  const request: AccountInfoRequest = {
+    command: 'account_info',
+    account,
+    ledger_index: 'current',
+  }
+  const data = await client.request(request)
+  return data.result.account_data.Sequence
+}
+
 /**
  * Sets the next valid sequence number for a transaction.
  *
@@ -236,14 +252,8 @@ export async function setNextValidSequenceNumber(
   client: Client,
   tx: Transaction,
 ): Promise<void> {
-  const request: AccountInfoRequest = {
-    command: 'account_info',
-    account: tx.Account,
-    ledger_index: 'current',
-  }
-  const data = await client.request(request)
   // eslint-disable-next-line no-param-reassign, require-atomic-updates -- param reassign is safe with no race condition
-  tx.Sequence = data.result.account_data.Sequence
+  tx.Sequence = await getNextValidSequenceNumber(client, tx.Account)
 }
 
 /**
@@ -270,16 +280,21 @@ async function fetchOwnerReserveFee(client: Client): Promise<BigNumber> {
  * @param client - The client object.
  * @param tx - The transaction object.
  * @param [signersCount=0] - The number of signers (default is 0). Only used for multisigning.
- * @returns A promise that resolves with void. Modifies the `tx` parameter to give it the calculated fee.
+ * @returns A promise that returns the fee.
  */
-export async function calculateFeePerTransactionType(
+
+async function calculateFeePerTransactionType(
   client: Client,
   tx: Transaction,
   signersCount = 0,
-): Promise<void> {
+): Promise<BigNumber> {
   const netFeeXRP = await getFeeXrp(client)
   const netFeeDrops = xrpToDrops(netFeeXRP)
   let baseFee = new BigNumber(netFeeDrops)
+
+  const isSpecialTxCost = ['AccountDelete', 'AMMCreate'].includes(
+    tx.TransactionType,
+  )
 
   // EscrowFinish Transaction with Fulfillment
   if (tx.TransactionType === 'EscrowFinish' && tx.Fulfillment != null) {
@@ -289,14 +304,18 @@ export async function calculateFeePerTransactionType(
       // eslint-disable-next-line @typescript-eslint/no-magic-numbers -- expected use of magic numbers
       scaleValue(netFeeDrops, 33 + fulfillmentBytesSize / 16),
     )
-  }
-
-  const isSpecialTxCost = ['AccountDelete', 'AMMCreate'].includes(
-    tx.TransactionType,
-  )
-
-  if (isSpecialTxCost) {
+  } else if (isSpecialTxCost) {
     baseFee = await fetchOwnerReserveFee(client)
+  } else if (tx.TransactionType === 'Batch') {
+    const rawTxFees = await tx.RawTransactions.reduce(async (acc, rawTxn) => {
+      const resolvedAcc = await acc
+      const fee = await calculateFeePerTransactionType(
+        client,
+        rawTxn.RawTransaction,
+      )
+      return BigNumber.sum(resolvedAcc, fee)
+    }, Promise.resolve(new BigNumber(0)))
+    baseFee = BigNumber.sum(baseFee.times(2), rawTxFees)
   }
 
   /*
@@ -304,7 +323,7 @@ export async function calculateFeePerTransactionType(
    * BaseFee × (1 + Number of Signatures Provided)
    */
   if (signersCount > 0) {
-    baseFee = BigNumber.sum(baseFee, scaleValue(netFeeDrops, 1 + signersCount))
+    baseFee = BigNumber.sum(baseFee, scaleValue(netFeeDrops, signersCount))
   }
 
   const maxFeeDrops = xrpToDrops(client.maxFeeXRP)
@@ -313,8 +332,25 @@ export async function calculateFeePerTransactionType(
     : BigNumber.min(baseFee, maxFeeDrops)
 
   // Round up baseFee and return it as a string
-  // eslint-disable-next-line no-param-reassign, @typescript-eslint/no-magic-numbers, require-atomic-updates -- safe reassignment.
-  tx.Fee = totalFee.dp(0, BigNumber.ROUND_CEIL).toString(10)
+  return totalFee.dp(0, BigNumber.ROUND_CEIL)
+}
+
+/**
+ * Calculates the fee per transaction type and sets it in the transaction.
+ *
+ * @param client - The client object.
+ * @param tx - The transaction object.
+ * @param [signersCount=0] - The number of signers (default is 0). Only used for multisigning.
+ * @returns A promise that resolves with void. Modifies the `tx` parameter to give it the calculated fee.
+ */
+export async function getTransactionFee(
+  client: Client,
+  tx: Transaction,
+  signersCount = 0,
+): Promise<void> {
+  const fee = await calculateFeePerTransactionType(client, tx, signersCount)
+  // eslint-disable-next-line @typescript-eslint/no-magic-numbers, require-atomic-updates, no-param-reassign -- fine here
+  tx.Fee = fee.toString(10)
 }
 
 /**
@@ -399,5 +435,73 @@ export function handleDeliverMax(tx: Payment): void {
 
     // eslint-disable-next-line no-param-reassign -- needed here
     delete tx.DeliverMax
+  }
+}
+
+/**
+ * Autofills all the relevant `x` fields.
+ *
+ * @param client - The client object.
+ * @param tx - The transaction object.
+ * @returns A promise that resolves with void if there are no blockers, or rejects with an XrplError if there are blockers.
+ */
+// eslint-disable-next-line complexity, max-lines-per-function -- needed here, lots to check
+export async function autofillBatchTxn(
+  client: Client,
+  tx: Batch,
+): Promise<void> {
+  const accountSequences: Record<string, number> = {}
+
+  for await (const rawTxn of tx.RawTransactions) {
+    const txn = rawTxn.RawTransaction
+
+    // Sequence processing
+    if (txn.Sequence == null && txn.TicketSequence == null) {
+      if (txn.Account in accountSequences) {
+        txn.Sequence = accountSequences[txn.Account]
+        accountSequences[txn.Account] += 1
+      } else {
+        const nextSequence = await getNextValidSequenceNumber(
+          client,
+          txn.Account,
+        )
+        const sequence =
+          txn.Account === tx.Account ? nextSequence + 1 : nextSequence
+        accountSequences[txn.Account] = sequence + 1
+        txn.Sequence = sequence
+      }
+    }
+
+    if (txn.Fee == null) {
+      txn.Fee = '0'
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- needed for JS checks
+    } else if (txn.Fee !== '0') {
+      throw new XrplError('Must have `Fee of "0" in inner Batch transaction.')
+    }
+
+    if (txn.SigningPubKey == null) {
+      txn.SigningPubKey = ''
+      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- needed for JS checks
+    } else if (txn.SigningPubKey !== '') {
+      throw new XrplError(
+        'Must have `SigningPubKey` of "" in inner Batch transaction.',
+      )
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- needed for JS checks
+    if (txn.TxnSignature != null) {
+      throw new XrplError(
+        'Must not have `TxnSignature` in inner Batch transaction.',
+      )
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- needed for JS checks
+    if (txn.Signers != null) {
+      throw new XrplError('Must not have `Signers` in inner Batch transaction.')
+    }
+
+    if (txn.NetworkID == null && txNeedsNetworkID(client)) {
+      txn.NetworkID = client.networkID
+    }
   }
 }