+++ packages/ripple-binary-codec/HISTORY.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 ### Added
-* Support for `Account Permissions` and `Account Permission Delegation` (XLS-74d, XLS-75d)
+* Support for `Account Permissions` and `Account Permission Delegation` (XLS-74d, XLS-75d).
+* Support for the `Batch` amendment (XLS-56).
 
 ### Fixed
 * add `MPTCurrency` support in `Issue` (rippled internal type)
@@ -18,7 +19,7 @@
 ## 2.2.0 (2024-12-23)
 
 ### Added
-* Support for the Multi-Purpose Token amendment (XLS-33)
+* Support for the Multi-Purpose Token amendment (XLS-33).
 
 ## 2.1.0 (2024-06-03)