+++ packages/xrpl/HISTORY.md
@@ -6,13 +6,17 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Support for `NFTokenMintOffer` (XLS-52)
+* Add Github Actions step to auto-generate documentation
 * Support for `Account Permissions` and `Account Permission Delegation` (XLS-74d, XLS-75d)
+* Add support for `Batch` amendment (XLS-56)
 
 ### Fixed
-* `OracleSet` transaction accepts hexadecimal string values for `AssetPrice` field
+* Fix `OracleSet` transaction to accept hexadecimal string values for `AssetPrice` field
 * `TransactionStream` model includes `hash` field in APIv2
 * `TransactionStream` model includes `close_time_iso` field only for APIv2
 * Adds `MPTCurrency` type
+* Better faucet support
+* Improve multisign fee calculations
 
 ## 4.2.0 (2025-2-13)